Applied "Some Mac Guy"'s Mac path changes.  His comments:
The code for the Initialize_Paths() function (which is for OS X only) requires that the maps, graphics, etc. folders be located in the resources directory of the application bundle if smw was compiled as an application bundle (the function does nothing if smw is run from the command line). While this is the best choice for most OS X applications, it makes it difficult for users to install fan-made content. 

It was also breaking on the level editor, since it was being called twice via convertPath. So I added a little bit to the top to make sure it returned immediately if it had already been run.